requirements.txt
aiohttp
beautifulsoup4
lxml
Pillow